refactor: pull arch out into separate file
